requirements.txt
opencv-python
numpy
pupil-apriltags
PyYAML

